chore: move sanity client to clients folder
