+++ web-registry/src/pages/CreateCreditClassInfo/CreateCreditClassInfo.tsx
@@ -14,7 +14,7 @@ import ResponsiveSlider from 'web-components/lib/components/sliders/ResponsiveSl
 import { Body, Title } from 'web-components/lib/components/typography';
 
 import { useAllCreateCreditClassPageQuery } from 'generated/sanity-graphql';
-import { client } from 'sanity';
+import { client } from 'lib/clients/sanity';
 
 import { WrappedImpactCard, WrappedResourcesCard } from 'components/atoms';
 import { HeroAction, HeroTitle, OverviewCards } from 'components/molecules';
