+++ web-registry/src/pages/LandStewards/LandStewards.tsx
@@ -18,7 +18,7 @@ import {
   TwoImageSection,
 } from '../../components/molecules';
 import { useAllLandStewardsPageQuery } from '../../generated/sanity-graphql';
-import { client } from '../../sanity';
+import { client } from '../../lib/clients/sanity';
 
 const useStyles = makeStyles()(theme => ({
   heroMain: {
